% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_annotated.R
\name{annotated}
\alias{annotated}
\title{For students: get annotations}
\usage{
annotated(...)
}
\arguments{
\item{...}{The call to function to get annotations.}
}
\description{
For students: get annotations
}
