% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_annotated.R
\name{create_annotation}
\alias{create_annotation}
\title{For instructors: create annotation}
\usage{
create_annotation(...)
}
\arguments{
\item{...}{An examplary call that produces a print output to be annotated.}
}
\description{
Create conditional annotations, and save them for further use.
}
\details{
Run create_annotation() and follow inline instructions, or add an unquoted call as an argument.
}
\examples{
 require(datasets)
 if(interactive())
   create_annotation(lm('dist ~ speed', cars))

}
