% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported_helpers.R
\name{parse_annotation_row}
\alias{parse_annotation_row}
\title{Helper to run inline r code (a lá .Rmd)}
\usage{
parse_annotation_row(rowtxt)
}
\arguments{
\item{rowtxt}{Annotation to parse.}
}
\description{
Helper to run inline r code (a lá .Rmd)
}
