% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_annotated.R
\name{annotated}
\alias{annotated}
\title{For students: get annotations}
\usage{
annotated(...)
}
\arguments{
\item{...}{The call to function to get annotations.}
}
\description{
For students: get annotations
}
\examples{
options(annotated.source="example.R") # set the source file or directory
m<-annotated(cfa('f1 =~ x1 + x2 + x3 + x4', HolzingerSwineford1939))

}
