% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_annotated.R
\name{annotated}
\alias{annotated}
\title{For learners: get annotations}
\usage{
annotated(...)
}
\arguments{
\item{...}{A call to a function to get annotations.}
}
\value{
The function invisibly returns an evaluated call to a function. For example, if `annotated(lm(a ~ b))` was called, it will return fitted `lm` object, as if annotated weren't there. Annotations are printed in the console at all times, even when the value is being assigned to an object.
}
\description{
For learners: get annotations
}
\examples{
if(interactive()) {
 options(annotated.source = "vignettes/lm.annot.R") # set the source file or directory
 require(datasets)
 lin.mod1 <- annotated(lm('dist ~ speed', cars))
}

}
