% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_annotated.R
\name{create_annotation}
\alias{create_annotation}
\title{For teachers: create annotation}
\usage{
create_annotation(...)
}
\arguments{
\item{...}{An examplary call that produces a print output to be annotated.}
}
\description{
For teachers: create annotation
}
\details{
run create_annotation() and follow the interactive instructions, or add an unquoted call.
}
\examples{
 require(datasets)
 create_annotation(lm('dist ~ speed', cars))

}
